client/requirements.txt
fastapi
uvicorn
langgraph
langchain-openai
python-dotenv
mcp
httpx